Added last-modified comparison after size check to determine what to sync.
Traditional rsync looks for equality in last-modified. Since we only have time uploaded to work with, this implementation opts to preserve the more recent copy, similar to rsync --update. I think this is relatively intuitive for uploads, but it could be problematic for downloads if the user is trying to "restore" as opposed to "get up to date" - I'm figuring in the case of the former it's easier to delete the local copy and try again than to recover inadvertently overwriting a file.

This also implements a --size-only option like rsync's which can overcome the potential pitfall above.